+++ wire-ios-request-strategy/Sources/Helpers/UserClient+QualifiedID.swift
@@ -16,24 +16,24 @@
 // along with this program. If not, see http://www.gnu.org/licenses/.
 //
 
-//import WireDataModel
-//
-//public extension UserClient {
-//
-//    /// An identifier build from the given properties of ``UserClient``. Returns `nil` if required properties are
-//    /// missing.
-//    var qualifiedClientID: QualifiedClientID? {
-//        guard
-//            let clientID = remoteIdentifier,
-//            let qualifiedID = user?.qualifiedID
-//        else {
-//            return nil
-//        }
-//
-//        return QualifiedClientID(
-//            userID: qualifiedID.uuid,
-//            domain: qualifiedID.domain,
-//            clientID: clientID
-//        )
-//    }
-//}
+import WireDataModel
+
+public extension UserClient {
+
+    /// An identifier build from the given properties of ``UserClient``. Returns `nil` if required properties are
+    /// missing.
+    var qualifiedClientID: QualifiedClientID? {
+        guard
+            let clientID = remoteIdentifier,
+            let qualifiedID = user?.qualifiedID
+        else {
+            return nil
+        }
+
+        return QualifiedClientID(
+            userID: qualifiedID.uuid,
+            domain: qualifiedID.domain,
+            clientID: clientID
+        )
+    }
+}
