+++ wire-ios-request-strategy/Sources/Request Strategies/Feature configurations/Payload/MLSFeatureProcessor.swift
@@ -38,8 +38,7 @@ public final class MLSFeatureProcessor: MLSFeatureProcessorProtocol {
     ) {
         self.mlsClientSyncManager = MLSClientSyncManager(
             coreCryptoProvider: coreCryptoProvider,
-            mlsService: mlsService
-        )
+            mlsService: mlsService)
         self.context = syncContext
     }
